Improve spectrum analyzer view. - paint optimization: using global gradient was not a good idea, performance-wise - add peaks in the spectrum - increase default FFT size to 2^15 - remove exception throwing in FourierTransform (even if they are very usefull, I want it the lightest possible for Bragi internal usage) - some naming and codestyle improvements in FFT code

+++ src/main/java/com/github/achaaab/bragi/gui/common/PainterThread.java
@@ -25,7 +25,7 @@ public class PainterThread extends Thread {
 	 * @param component component to repaint
 	 * @param fps       required number of frames per second
 	 */
-	public PainterThread(JComponent component, int fps) {
+	public PainterThread(JComponent component, double fps) {
 
 		this.component = component;
